refactor: change model file location into each features
