+++ app/(server)/_features/room/model.ts
