+++ app/(server)/_schema/index.ts
@@ -1,3 +1,3 @@
-import {rooms} from './rooms'
+import {rooms} from '../_features/room/model'
 
 export default {rooms}
